Issue 128 - RPC::Agent.meta is broken
Since in the DSL style agents we do metadata not in the
instance of the class but rather at the singleton class
level we were creating @@meta which was bad since all the
agents shared the same @@meta.

This didnt cause any problems but the data wasnt correctly
stored.

This keeps backward compatability by letting agents still modify
meta data during startup_hook but when they use the new format
we replace the default getter for @meta with one that
always returns the metadata supplied data.  Once this is done the
setter will work without throwing exceptions but  it wont actually
do anything

This is not an ideal fix, we should further refactor the RPC::Agent
so it's a cleaner DSL.