Issue 143 - Any repeating exception during registration causes a CPU loop
In cases where sending registration would fail with an exception and when
those exceptions would happen on every registration run there would be a
loop that consumed a lot of resources.

This happens because we only slept for the configured period after
successful registration events and not failing ones.  This patch add
a sleep after failed registration events.