5084 - fact checking should support common operators: <=, >=, =~, ...
Rework the logic in has_fact? to be more rubylike and contain the
casting of values to only the cases where it's needed rather than
do it all the time.

Put the fact operator parsing into a reusable method and apply to
the -W / --with parsing as well and treat => like >= and =< like <=