7574 - Single app framework needs better error reporting for config file errors
In the past any error in loading config or parsing application list would simply
result in an empty applicaiton list being returned.  In cases of a configuration
error this masked the real cause and made debugging really hard.

We now print the exception received instead and exit with a exit code of 1 which
should improve the situation marginally.

Also improve the test coverage for the Applications class