7583 - Plugin Activation - Configuration and Supportability Checks
Agents can now have an class method called activate? that can
return true, false or raise an exception.  For false or exception
the agent will not be added to the running mcollectived and so
the node will not respond to discovery.

The SimpleRPC base agent has a default activate? method that
looks in the config file and allows users to disable agents
based on config.

A helper activate_when was added to the SimpleRPC base agent
that users can use to override the default activate? method
with their own logic for activation.

This should be used by agents to figure out if they're capable
of working on a given node.  It should make deployment simpler
as given correctly written agents you should now be able to
copy all agents to all machines.

Test coverage was added to the M::Agents class