7620 - Destination names used during subscribe and unsubscribe should be created in the connectors
MCollective core now does not need to know how the middleware works
all it does is ask the connector to subscribe to either a :broadcast,
:directed or :reply queue.  The connector then decides how to handle
that.  Further work in #7619 is needed to fully realize the abstraction.

The Stomp connector was updated to support the new scheme.  Backwards
compatability was maintained.

Test coverage was added for the Stomp connector