7225 - mcollectived should be able to subscribe to a per identity queue
Add a new config option direct_addressing that will cause the main runner
to request :directed subscriptions from the middleware plugins.

This setting is off by default since the default stomp plugin really
isn't going to be able to do really intelligent stuff with this and
on big networks it will have big impacts on middleware.  The proposed
ActiveMQ specific plugin will be able to do this in a way that does
not have such a huge impact on the middleware.

For Stomp we've added a new config option queueprefix that compliments
topicprefix and the plugin will simply subscribe to a queue that has a
MD5 of the configured identity.  We're using the md5 as thats the safest
way to ensure there's no illegal characters in the resulting queue names.

The intention with this ability is so that we can address individual nodes
directly without broadcasting.  The use cases for this are - among others:

- Web apps where people will be clicking arbitrary hosts, the library will
  fall back to direct mode in that case
- For calls to M::RPC::Client#custom_request where the list of expected
  replies are some arbitrary low number like 20.  In a big network talking
  to those hosts directly will have a lower impact than the broadcast
  mode
- Add the potential to do batched requests.  With the broadcast mode
  it's a bit all or nothing unless restrictive filters are generated.
  By combining the broadcast for discovery with directed requests in
  groups of n hosts we can slowly go through a network rolling out
  change in managable batches.
- This will enable truely pluggable discovery.  External resources
  like BPM systems can supply a list of hosts they wish to talk to
  and we can do that easily.
- Queued requests for off-line machines

None of these abilities are in the libraries now, this is just the
enabling architecture.