8280 - RPC::Helpers#rpcresults should support JSON
Add the ability for the printrpc helper to return JSON data.
Any application that is using the printrpc helper to display
data will now be able to just add -j to the command line and
return JSON instead.

If STDOUT is a tty the JSON will be pretty formatted else it
will just be a big blob of machine friendly JSON.

The basic use case is to eventually allow piping RPC results
from one query into the following query - perhaps with some
grep like filtering - thus avoiding discovery and allowing
totally arbitrary actions on arbitrary sets of hosts.

This also adjusts the progress bar and the stats to only display
if STDOUT is a tty.