8419 - Filters with invalid syntax do not fail, possible unintentional execution in environment.
If users supplied fact filters on the CLI or in ruby code that did not
match our expected valid fact formats the effect would be as if the
filter was never specified.

This would result in requests being run on the wrong machines.  We
now raise an exception rather than just returning false