8325 - A default TTL should be supported
As messages are now possibly sent to queues and can hang around the
middleware for a very long time we need to support a TTL.  Nodes that
go off-line between discovery and request being sent may come back
a month later and find these old messages sitting there and then take
the actions which would not be desirable.

Requests will now get a default TTL of 60 seconds and during validation
of a message any messages older than this will not validate.

The various security plugins will accept a TTL when encoding the messages
and default to 60 if not given, similarly the Message object will default
to 60 TTL.

As a side effect security plugins now need a TTL when encoding requests
which means we now need to upgrade all security plugins in use when going
to a version running this version.

Future tickets will allow the TTL to be set per message.