10671 - Security plugins should not decrypt messages for other clients
Create a helper function on the MC::Security::Base that security plugins
can use to decide if they should do all the hard work.

The basic structure of a security plugins decode phase is:

  msg = deserialize(payload)

  decrypt or decode(msg)

the 2nd step is generally the really expensive step where SSL or RSA
comes into play, these plugins can now use the should_process_msg?
helper to raise an appropriate exception if the message is not for us
before doing the expensive work