6275 - mc-* in $PATH to `mc` confuses the unified-binary
The mco command supports creating a symlink called mco-foo to it which
would invoke the foo application.

The regular expression thats used to drive the logic was not strict
enough and so if the mco script was in a directory called mc-anything it
would fail.

This makes the regular expression more strict reducing the cases where
it will fail.  There is still a few edge cases but this feature is
almost never used now that migration to the unified binary is complete
so it's not worth spending much time on.