12703 - stomp and activemq connectors should handle Stomp::Error::DuplicateSubscription
Stomp 1.2.0 introduce new behavior where duplicate subscriptions are
suppressed, this is good we do not want to duplicate subscriptions but
we should now handle the exceptions correctly

With the ActiveMQ plugin we do subscriptions that include headers, so
that one client/agent can subscribe to the same queue multiple times
but just with different sets of headers supplied.

The Stomp gem does not consider headers when tracking duplicate
subscriptions so this would have been impossible.  So we now come up
with our own unique subscription names in those cases and supply it to
the stomp gem to improve its internal tracking