12719 - DDL returns {} for unknown action interfaces but the RPC application does not handle this
When an action gets called with paramters we go through each param and
convert them where possible into the formats the ddl expect.  So the
string "1" when given on the CLI gets turned into the number 1.0 if the
DDL expects a float.

The method that did this did not handle the case where the action does
not exist and would raise a nilClass error, this adds that check