12854 - Users should be able to disable and enable batch mode during the life of a client
The batch mode behavior was either on or off for the duration of the
client, if a user supplied a batch mode on the command line there was no
way for the program to decide that some calls should be subject to batch
mode and others should not.

Additionally once enabled batch mode could not be disabled again.

This commit allow the batch_size to be set to 0 to disable the batch
mode for the client.  This is allowed either for 1 single rpc call by
passing :batch_size to the rpc call as an argument or by setting
batch_size to 0 on the client object

Also adds a bunch of test coverage to previously untested code