12927 - Application#application_failure should not handle SystemExit
Users can use exit() in applications as is expected in any normal ruby
script, the application_failure handler though gets exceptions and
prints backtraces.

In some cases the application class caught SystemExit and did not pass
those to the application_failure but it was only done in some cases.

With this commit the appliction_failure method now handles SystemExit
correctly so that any SystemExit gets processed through this rather than
just the few places where we had the specific handling

Also adds test coverage to code that did not have it before