Feature #12597
Implementation of the mcollective plugin packaging application.

This implementation uses the mco plugin application to identify a plugin structure from a commond line argument or
by parsing its directory structure. The identified plugin type is then created containing the list of packages to
be created from this plugin, which files are contained in the plugin as well as other metadata. Plugable packagers
are then used to create appropriate output, based on osfamily values via facter or commond line arguments.