13020 - Handle Stomp 1.2.0 Stomp::Error::NoCurrentConnection better
While adding Stomp 1.1 support to the Stomp gem upstream tightened up
various connection handling issues in the gem.  It now raises an
exception when you try to receive on a disconnected socket so we're
handling that on our side.

For gem 1.2.0 there will be a complete lack of logging when this happens
but thats a bug in the Stomp Gem, I've also sent a pull request upstream
to address that issue.