13064 - implemented_by scripts should have a conventional place
We support actions in external commands but do not provide a canonical
place to put them.

Having a known convenstion means the plugin packager can be programmed
to deal with these files automatically.

After this commit if you specified a implemented_by script like:

   implemented_by "foo.sh"

The script will be searched for in all the libdir in
libdir/agents/agent_name/foo.sh till it finds it. If a full path is
supplied the full path will be used instead.